+++ frontend/src/app/(page)/(auth)/layout.tsx
@@ -1,4 +1,4 @@
-import ReturntoHome from "@/_Components/cross-componets/returnto-home";
+import ReturntoHome from "@/app/_Components/cross-componets/returnto-home";
 
 
 
@@ -9,7 +9,6 @@ export default function RootLayout({
     children: React.ReactNode;
 }>) {
     return (
-
         <div className="flex w-full min-h-screen bg-gray-50 dark:bg-gray-900 content-transition">
             <ReturntoHome />
             {children}
