+++ frontend/src/app/(page)/(auth)/login/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { LoginForm } from '@/_Components/login/login-form'
+import { LoginForm } from '@/app/_Components/login/login-form'
 import React from 'react'
 
 function page() {
