+++ frontend/src/app/(page)/(main)/product/[name]/page.tsx
