+++ frontend/src/app/_Components/main/Hero.tsx
