+++ frontend/src/app/_Components/main/seasonal-section.tsx
