+++ frontend/src/app/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import './globals.css'
+import './styles/globals.css'
 import { QueryProviders } from "./provider/query-client-provider";
 
 export const metadata: Metadata = {
