+++ frontend/src/app/styles/globals.css
